requirements.txt
matplotlib==3.6.0
natsort==8.3.1
numpy==1.21.5
opencv_python_headless==4.7.0.68
pandas==1.3.5
Pillow==9.4.0
pytorch_lightning==2.0.0
scikit_learn==1.2.2
timm==0.6.12
torch==1.12.0+cu113
torchvision==0.13.0+cu113
tqdm==4.63.0
